separate scopes classes in different files
