+++ lib/carto/oauth_provider/scopes/category.rb
@@ -1,0 +1,14 @@
+module Carto
+  module OauthProvider
+    module Scopes
+      class Category
+        attr_reader :description, :icon
+
+        def initialize(description, icon = nil)
+          @description = description
+          @icon = icon
+        end
+      end
+    end
+  end
+end
